requirements.txt
aiohttp                   3.7.4.post0
altgraph                  0.17.2
APScheduler               3.6.3
async-timeout             3.0.1
attrs                     21.4.0
backports.zoneinfo        0.2.1
beautifulsoup4            4.11.1
cachetools                4.2.2
certifi                   2021.10.8
chardet                   4.0.0
charset-normalizer        2.0.12
click                     7.1.2
discord.py                1.7.3
Flask                     2.0.3
Flask-Login               0.5.0
flask-ngrok               0.0.25
Flask-WTF                 1.0.0
future                    0.18.2
greenlet                  1.1.2
idna                      3.3
itsdangerous              2.1.0
Jinja2                    3.0.3
MarkupSafe                2.1.0
multidict                 6.0.2
pefile                    2021.9.3
Pillow                    9.0.1
pip                       22.0.3
pyinstaller               4.10
pyinstaller-hooks-contrib 2022.2
PyQt5                     5.15.4
pyqt5-plugins             5.15.4.2.2
PyQt5-Qt5                 5.15.2
PyQt5-sip                 12.9.1
pyqt5-tools               5.15.4.3.2
python-dotenv             0.19.2
python-telegram-bot       13.11
pytz                      2022.1
pytz-deprecation-shim     0.1.0.post0
pywin32-ctypes            0.2.0
qt5-applications          5.15.2.2.2
qt5-tools                 5.15.2.1.2
requests                  2.27.1
setuptools                49.2.1
six                       1.16.0
soupsieve                 2.3.2.post1
SQLAlchemy                1.4.32
tornado                   6.1
typing_extensions         4.1.1
tzdata                    2022.1
tzlocal                   4.1
urllib3                   1.26.8
vk-api                    11.9.7
Werkzeug                  2.0.3
wikipedia                 1.4.0
WTForms                   3.0.1
yarl                      1.7.2